Save latest updates which moves all DeviceMedia asset display into a separate component.
